MAJ par-rapport a Caml Light 0.7
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@182 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/unix/socketaddr.c
@@ -10,7 +10,7 @@
 #include "socketaddr.h"
 
 value alloc_inet_addr(a)
-     unsigned long a;
+     unsigned int a;
 {
   value res;
   res = alloc(1, Abstract_tag);
